+++ playground/src/router/routes/misc.ts
@@ -1,0 +1,12 @@
+export const miscRoutes = [
+  {
+    path: '/misc',
+    name: 'Text 3D',
+    component: () => import('../../pages/misc/Text3DDemo.vue'),
+  },
+  {
+    path: '/misc/directives',
+    name: 'Directives',
+    component: () => import('../../pages/misc/Directives.vue'),
+  },
+]