+++ src/directives/index.ts
@@ -1,7 +1,5 @@
 import { vLog } from './vLog'
 import { vLightHelper } from './vLightHelper'
-import { vAlwaysLookAt } from './vAlwaysLookAt'
 import { vDistanceTo } from './vDistanceTo'
-import { vRotate } from './vRotate'
 
-export { vLog, vLightHelper, vAlwaysLookAt, vDistanceTo, vRotate }
+export { vLog, vLightHelper, vDistanceTo }